+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Improvements
 - Support range requests when downloading DICOMweb files ([#1444](../../pull/1444))
 - Bypass some scaling code when compositing multi sources ([#1447](../../pull/1447), [#1449](../../pull/1449))
+- Do not create needless alpha bands in the multi source ([#1451](../../pull/1451))
+- Infer DICOM file size, when possible ([#1448](../../pull/1448))
 
 ## 1.27.1